requirements.txt
pandas
numpy
fyers_api
API credentials - API secret key and client id
schedule